refactor: enum class
