+++ packages/near-api-js/src/utils/enums.ts
@@ -1,14 +1,1 @@
-/** @hidden @module */
-export abstract class Enum {
-    enum: string;
-
-    constructor(properties: any) {
-        if (Object.keys(properties).length !== 1) {
-            throw new Error('Enum can only take single value');
-        }
-        Object.keys(properties).map((key: string) => {
-            (this as any)[key] = properties[key];
-            this.enum = key;
-        });
-    }
-}
+export { Enum } from '@near-js/types';
