+++ packages/transactions/src/actions.ts
@@ -1,22 +1,8 @@
 import { PublicKey } from '@near-js/crypto';
+import { Enum } from '@near-js/types';
 
 import { DelegateAction } from './delegate';
 import { Signature } from './signature';
-
-// TODO determine why subclassing is still necessary even though `enum`
-//  cannot be set in the base class or it will not be borsh-serializable
-abstract class Enum {
-    abstract enum: string;
-
-    constructor(properties: any) {
-        if (Object.keys(properties).length !== 1) {
-            throw new Error('Enum can only take single value');
-        }
-        Object.keys(properties).map((key: string) => {
-            (this as any)[key] = properties[key];
-        });
-    }
-}
 
 export class FunctionCallPermission {
     allowance?: bigint;
