+++ packages/types/src/index.ts
@@ -1,2 +1,3 @@
+export * from './enum';
 export * from './errors';
 export * from './provider';
